feat: add login and register events
Tracking won't be perfect on these, but it's as close as we can
really get when only using the frontend.

Login tracks on a successful sign in from `app.session.login`,
while sign up waits for a successful submission of the SignUpModal
but this means other auth sources, if any, will not be counted.